pagination + a bit better board pages + working previews !!!!!!!!
